docs(README.md): update README with author info and additional folder descriptions
Add author information for Le Chen and affiliation with Auburn University.
Clarify the source file type for the slides as LaTeX. Introduce a new
section detailing the purpose and contents of various folders used in
the project, such as `exp13`, `tetromino`, and `figs`. This enhances
the documentation by providing more context and guidance for users
navigating the project structure, making it easier to understand the
organization and purpose of different assets and configurations.